backend/requirements.txt
fastapi==0.115.5
python-dotenv==1.2.1
uvicorn[standard]==0.32.1
python-multipart==0.0.17
openai==1.55.3
PyPDF2==3.0.1
python-docx==1.1.2
beautifulsoup4==4.12.3
openpyxl==3.1.5
